requirements.txt
streamlit>=1.37.0
pydantic>=2.8.0
python-dotenv>=1.0.1
langchain>=0.2.12
tiktoken>=0.7.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
plotly>=5.22.0
pytest>=8.2.0
